Packages content in VnV with tests
