+++ src/app/validation-and-verification/vnv-packages-detail/vnv-packages-detail.component.html
@@ -18,12 +18,16 @@
 </div>
 
 <form class="detail">
-  <mat-form-field class="vendor" *ngIf="detail['vendor']">
+  <mat-form-field *ngIf="detail['vendor']">
     <input matInput placeholder="Vendor" value="{{ detail['vendor'] }}" disabled>
   </mat-form-field>
 
-  <mat-form-field class="type" *ngIf="detail['type']">
+  <mat-form-field *ngIf="detail['type']">
     <input matInput placeholder="Type" value="{{ detail['type'] }}" disabled>
+  </mat-form-field>
+
+  <mat-form-field *ngIf="detail['status']">
+    <input matInput placeholder="Status" value="{{ detail['status'] }}" disabled>
   </mat-form-field>
 
   <mat-form-field class="updatedAt" *ngIf="detail['updatedAt']">
@@ -60,6 +64,32 @@
     <span class="sectionTitle">VNF</span>
     <span class="message" *ngIf="detail['vnf'].length == 0">There are no virtual network functions to display</span>
     <table mat-table *ngIf="detail['vnf'].length > 0" [dataSource]="detail['vnf']">
+      <ng-container matColumnDef="vendor">
+        <th mat-header-cell *matHeaderCellDef> Vendor </th>
+        <td mat-cell *matCellDef="let element"> {{ element.vendor }} </td>
+      </ng-container>
+
+      <ng-container matColumnDef="name">
+        <th mat-header-cell *matHeaderCellDef> Name </th>
+        <td mat-cell *matCellDef="let element"> {{ element.name }} </td>
+      </ng-container>
+
+      <ng-container matColumnDef="version">
+        <th mat-header-cell *matHeaderCellDef> Version </th>
+        <td mat-cell *matCellDef="let element"> {{ element.version }} </td>
+      </ng-container>
+
+      <tr mat-header-row *matHeaderRowDef="displayedColumns">
+      </tr>
+      <tr mat-row *matRowDef="let row; columns: displayedColumns;">
+      </tr>
+    </table>
+  </div>
+
+  <div class="tests" *ngIf="detail['tests']">
+    <span class="sectionTitle">Tests</span>
+    <span class="message" *ngIf="detail['tests'].length == 0">There are no tests to display</span>
+    <table mat-table *ngIf="detail['tests'].length > 0" [dataSource]="detail['tests']">
       <ng-container matColumnDef="vendor">
         <th mat-header-cell *matHeaderCellDef> Vendor </th>
         <td mat-cell *matCellDef="let element"> {{ element.vendor }} </td>
