+++ src/app/validation-and-verification/vnv-packages-detail/vnv-packages-detail.component.ts
@@ -1,22 +1,20 @@
 import { Component, OnInit, ViewEncapsulation } from "@angular/core";
 import { Router, ActivatedRoute } from "@angular/router";
 
-import { ServicePlatformService } from "../service-platform.service";
 import { CommonService } from "../../shared/services/common/common.service";
 
 @Component({
-  selector: "app-sp-packages-detail",
-  templateUrl: "./sp-packages-detail.component.html",
-  styleUrls: ["./sp-packages-detail.component.scss"],
+  selector: "app-vnv-packages-detail",
+  templateUrl: "./vnv-packages-detail.component.html",
+  styleUrls: ["./vnv-packages-detail.component.scss"],
   encapsulation: ViewEncapsulation.None
 })
-export class SpPackagesDetailComponent implements OnInit {
+export class VnvPackagesDetailComponent implements OnInit {
   displayedColumns = ["name", "vendor", "version"];
   loading: boolean;
   detail = {};
 
   constructor(
-    private servicePlatformService: ServicePlatformService,
     private commonService: CommonService,
     private router: Router,
     private route: ActivatedRoute
@@ -39,7 +37,7 @@ export class SpPackagesDetailComponent implements OnInit {
     this.loading = true;
 
     this.commonService
-      .getOnePackage("sp", uuid)
+      .getOnePackage("vnv", uuid)
       .then(response => {
         this.loading = false;
         this.detail = response;
@@ -50,6 +48,9 @@ export class SpPackagesDetailComponent implements OnInit {
         if (this.detail["vnf"].lenght < 1) {
           this.detail["vnf"] = [];
         }
+        if (this.detail["tests"].lenght < 1) {
+          this.detail["tests"] = [];
+        }
       })
       .catch(err => {
         this.loading = false;
@@ -59,6 +60,6 @@ export class SpPackagesDetailComponent implements OnInit {
   }
 
   close() {
-    this.router.navigate(["service-platform/packages"]);
+    this.router.navigate(["../"], { relativeTo: this.route });
   }
 }
