+++ src/app/validation-and-verification/vnv-packages/vnv-packages.component.scss
@@ -13,12 +13,7 @@ app-vnv-packages {
     width: 95%;
   }
 
-  td.mat-cell {
-    cursor: auto;
-  }
-
   tr.mat-row {
-    border-bottom: none;
     background-image: linear-gradient(
       to right,
       $basic-background-color 4%,
@@ -27,8 +22,6 @@ app-vnv-packages {
       $basic-background-color 98%,
       $basic-background-color 100%
     );
-    align-items: baseline;
-    min-height: 29px;
   }
 
   .mat-cell:nth-child(1),
