+++ src/app/validation-and-verification/vnv-packages/vnv-packages.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, ViewEncapsulation } from "@angular/core";
-import { ActivatedRoute } from "@angular/router";
+import { ActivatedRoute, Router } from "@angular/router";
 
 import { CommonService } from "../../shared/services/common/common.service";
 
@@ -17,6 +17,7 @@ export class VnvPackagesComponent implements OnInit {
 
   constructor(
     private commonService: CommonService,
+    private router: Router,
     private route: ActivatedRoute
   ) {}
 
@@ -51,5 +52,9 @@ export class VnvPackagesComponent implements OnInit {
         this.loading = false;
         this.commonService.openSnackBar(err, "");
       });
+  }
+
+  openPackage(row) {
+    this.router.navigate([row.uuid], { relativeTo: this.route });
   }
 }
